requirements.txt
alembic==1.3.1
APScheduler==3.6.3
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.2
Click==7.0
cryptography==2.8
feedparser==5.2.1
Flask==1.1.1
Flask-DotEnv==0.1.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.2.0
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
tornado==6.0.3
tzlocal==2.0.0
Werkzeug==0.16.0
